Use pop overs in proteom evidence
